+++ packages/create-widget/views/preact/template/src/client.js
@@ -1,4 +1,4 @@
-import { render } from 'preact';
+import { render, hydrate } from 'preact';
 import { unmountComponentAtNode } from 'preact/compat';
 import { createMerkurWidget, createMerkur } from '@merkur/core';
 import { widgetProperties } from './widget';
@@ -10,11 +10,16 @@ function createWidget(widgetParams) {
     ...widgetParams,
     $dependencies: {
       render,
+      hydrate,
       unmountComponentAtNode,
     },
     mount(widget) {
       const View = widget.View();
       const container = document.querySelector(widget.props.containerSelector);
+
+      if (container && container.children.length) {
+        return widget.$dependencies.hydrate(View, container);
+      }
 
       return widget.$dependencies.render(View, container);
     },