+++ packages/integration-react/src/MerkurComponent.jsx
@@ -1,16 +1,13 @@
+import React from 'react';
 import { getMerkur } from '@merkur/core';
 import { loadScriptAssets, loadStyleAssets } from '@merkur/integration';
-import React from 'react';
 
 // error event name from @merkur/plugin-error
 const MERKUR_ERROR_EVENT_NAME = '@merkur/plugin-error.error';
 
-function WidgetWrapper({ html, widgetClassName }) {
+function WidgetWrapper({ html, className }) {
   return (
-    <div
-      className={widgetClassName}
-      dangerouslySetInnerHTML={{ __html: html }}
-    />
+    <div className={className} dangerouslySetInnerHTML={{ __html: html }} />
   );
 }
 
@@ -26,17 +23,24 @@ export default class MerkurComponent extends React.Component {
 
     this.state = {
       encounteredError: false,
-      loaded: false,
+      assetsLoaded: false,
     };
   }
 
-  shouldComponentUpdate(nextProps) {
-    if (!this.props.widgetProperties || !this._widget) {
+  shouldComponentUpdate(nextProps, nextState) {
+    if (
+      !this.props.widgetProperties ||
+      this.state.assetsLoaded !== nextState.assetsLoaded ||
+      this.state.encounteredError !== nextState.encounteredError
+    ) {
       return true;
     }
 
-    // TODO refactoring
-    if (nextProps.widgetProperties && nextProps.widgetProperties.props) {
+    if (
+      this._widget &&
+      nextProps.widgetProperties &&
+      nextProps.widgetProperties.props
+    ) {
       for (let key in nextProps.widgetProperties.props) {
         if (
           !this.props.widgetProperties ||
@@ -45,11 +49,17 @@ export default class MerkurComponent extends React.Component {
             this.props.widgetProperties.props[key]
         ) {
           this._widget.setProps(nextProps.widgetProperties.props);
+
+          return false;
         }
       }
     }
 
-    if (nextProps.widgetProperties && nextProps.widgetProperties.state) {
+    if (
+      this._widget &&
+      nextProps.widgetProperties &&
+      nextProps.widgetProperties.state
+    ) {
       for (let key in nextProps.widgetProperties.state) {
         if (
           !this.props.widgetProperties ||
@@ -58,6 +68,8 @@ export default class MerkurComponent extends React.Component {
             this.props.widgetProperties.state[key]
         ) {
           this._widget.setState(nextProps.widgetProperties.state);
+
+          return false;
         }
       }
     }
@@ -67,24 +79,32 @@ export default class MerkurComponent extends React.Component {
 
   componentDidMount() {
     this._isMounted = true;
-    this._tryCreateWidget();
+    this._loadWidgetAssets();
   }
 
-  componentDidUpdate(prevProps) {
+  componentDidUpdate(prevProps, prevState) {
     const { widgetProperties: currentWidgetProperties } = this.props;
     const { widgetProperties: prevWidgetProperties } = prevProps;
+
+    // Mount widget after resources have been loaded
+    if (
+      this.state.assetsLoaded &&
+      prevState.assetsLoaded !== this.state.assetsLoaded
+    ) {
+      return this._mountWidget();
+    }
 
     if (!currentWidgetProperties && prevWidgetProperties) {
       this._removeWidget();
       this.setState({
-        loaded: false,
+        assetsLoaded: false,
       });
 
       return;
     }
 
     if (currentWidgetProperties && !prevWidgetProperties) {
-      return this._tryCreateWidget();
+      return this._loadWidgetAssets();
     }
 
     if (!currentWidgetProperties && !prevWidgetProperties) {
@@ -98,10 +118,10 @@ export default class MerkurComponent extends React.Component {
       this._removeWidget();
       this.setState(
         {
-          loaded: false,
+          assetsLoaded: false,
         },
         () => {
-          this._tryCreateWidget();
+          this._loadWidgetAssets();
         }
       );
     }
@@ -113,20 +133,22 @@ export default class MerkurComponent extends React.Component {
 
   render() {
     const { widgetProperties, widgetClassName } = this.props;
-    const { encounteredError, loaded } = this.state;
+    const { encounteredError, assetsLoaded } = this.state;
 
-    if (!widgetProperties || encounteredError) {
+    if (
+      !widgetProperties ||
+      encounteredError ||
+      (this._isClient() && !this._isSSRHydrate() && !assetsLoaded)
+    ) {
       return this._renderFallback();
     }
 
     const html = this._getWidgetHTML();
-    if (this._isClient() && !loaded && !(!this._isMounted && html)) {
-      return this._renderPlaceholder();
-    }
 
     return (
       <>
-        {this._renderStyleAssets()}
+        {(!this._isClient() || this._isSSRHydrate()) &&
+          this._renderStyleAssets()}
         <WidgetWrapper className={widgetClassName} html={html} />
       </>
     );
@@ -145,28 +167,12 @@ export default class MerkurComponent extends React.Component {
     return null;
   }
 
-  _renderPlaceholder() {
-    const { widgetProperties, widgetClassName } = this.props;
-    const { placeholder } = widgetProperties;
-
-    if (!placeholder) {
-      return null;
-    }
-
-    return <WidgetWrapper className={widgetClassName} html={placeholder} />;
-  }
-
   _renderStyleAssets() {
     const { widgetProperties } = this.props;
-    if (!widgetProperties || !Array.isArray(widgetProperties.assets)) {
-      return null;
-    }
+    const assets =
+      (Array.isArray(widgetProperties.assets) && widgetProperties.assets) || [];
 
-    if (this._isClient() || !this._getWidgetHTML()) {
-      return null;
-    }
-
-    return widgetProperties.assets.map((asset, key) => {
+    return assets.map((asset, key) => {
       switch (asset.type) {
         case 'stylesheet':
           return <link rel="stylesheet" href={asset.source} key={key} />;
@@ -187,17 +193,7 @@ export default class MerkurComponent extends React.Component {
       return this._html;
     }
 
-    this._html = this.props.widgetProperties.html || '';
-
-    if (typeof document !== 'undefined') {
-      const container = document.querySelector(
-        this.props.widgetProperties.props.containerSelector
-      );
-
-      if (container && container.children && container.children[0]) {
-        this._html = container.children[0].outerHTML;
-      }
-    }
+    this._html = this.props.widgetProperties.html || this._getSSRHtml();
 
     return this._html;
   }
@@ -229,52 +225,100 @@ export default class MerkurComponent extends React.Component {
     this._html = null;
   }
 
-  _tryCreateWidget() {
-    const { widgetProperties, onWidgetMounted, debug } = this.props;
+  /**
+   * Loads widget assets into page.
+   */
+  _loadWidgetAssets() {
+    const { widgetProperties } = this.props;
 
     if (!widgetProperties || this._widget) {
       return;
     }
 
-    Promise.all([
-      this._getLoadStyleAssetsPromise(widgetProperties.assets),
+    return Promise.all([
+      loadStyleAssets(widgetProperties.assets),
       loadScriptAssets(widgetProperties.assets),
     ])
-      .catch((error) => this._handleError(error))
-      .then(async () => {
-        const merkur = getMerkur();
-        this._widget = await merkur.create(widgetProperties);
-        await this._widget.mount();
-
-        if (typeof this._widget.on === 'function') {
-          // widget might not be using @merkur/plugin-event-emitter
-          this._widget.on(MERKUR_ERROR_EVENT_NAME, this._handleClientError);
-        }
-
-        if (typeof onWidgetMounted === 'function') {
-          onWidgetMounted(this._widget);
-        }
-      })
-      .catch((error) => {
-        if (debug) {
-          console.warn(error);
-        }
-      });
+      .then(
+        () =>
+          new Promise((resolve) => {
+            this.setState(
+              {
+                assetsLoaded: true,
+              },
+              () => {
+                resolve();
+              }
+            );
+          })
+      )
+      .catch((error) => this._handleError(error));
   }
 
-  _getLoadStyleAssetsPromise(assets) {
-    return loadStyleAssets(assets).then(
-      new Promise((resolve) => {
-        this.setState(
-          {
-            loaded: true,
-          },
-          () => {
-            resolve();
-          }
-        );
-      })
-    );
+  /**
+   * Creates and mounts widget instance after all resource loaded.
+   */
+  async _mountWidget() {
+    const { widgetProperties, onWidgetMounted, debug = false } = this.props;
+
+    if (!widgetProperties || this._widget) {
+      return;
+    }
+
+    try {
+      const merkur = getMerkur();
+      this._widget = await merkur.create(widgetProperties);
+      await this._widget.mount();
+
+      if (typeof this._widget.on === 'function') {
+        // widget might not be using @merkur/plugin-event-emitter
+        this._widget.on(MERKUR_ERROR_EVENT_NAME, this._handleClientError);
+      }
+
+      if (typeof onWidgetMounted === 'function') {
+        onWidgetMounted(this._widget);
+      }
+    } catch (error) {
+      if (debug) {
+        console.warn(error);
+      }
+    }
+  }
+
+  /**
+   * Returns server-side rendered html, if its the first render on client
+   * after SSR.
+   *
+   * @return {string} server-side rendered html, if it's not available, returns empty string.
+   */
+  _getSSRHtml() {
+    if (
+      !this._isMounted &&
+      this._isClient() &&
+      typeof document !== 'undefined'
+    ) {
+      const container = document.querySelector(
+        this.props.widgetProperties.props.containerSelector
+      );
+
+      return (
+        container &&
+        container.children &&
+        container.children[0] &&
+        container.children[0].outerHTML
+      );
+    }
+
+    return '';
+  }
+
+  /**
+   * Checks if it's the first render after SSR.
+   *
+   * @return {boolean} true in case of a first render after SSR, otherwise false.
+   */
+  _isSSRHydrate() {
+    return this._getSSRHtml().length > 0;
   }
 
   _isClient() {